as.envlist {devtools}	R Documentation

Convert a list of environments to an envlist object.

Description

Convert a list of environments to an envlist object.

Usage

  as.envlist(x)

Arguments

x
A list of environments.

[Package devtools version 1.3 Index]