Build package vignettes.

Builds package vignettes using the same algorithm that R CMD build does. This means including non-Sweave vignettes, using makefiles (if present), and copying over extra files. You need to ensure that these files are not included in the built package - ideally they should not be checked into source, or at least excluded with .Rbuildignore

build_vignettes(pkg = ".", dependencies = "VignetteBuilder", quiet = TRUE)

Arguments

pkg
package description, can be path or package name. See as.package for more information

dependencies
logical indicating to also install uninstalled packages which this pkg depends on/links to/suggests. See argument dependencies of install.packages.

quiet
If TRUE, suppresses most output. Set to FALSE if you need to debug.

See also