build_win {devtools}	R Documentation

Build windows binary package.

Description

This function works by bundling source package, and then uploading to http://win-builder.r-project.org/. Once building is complete you'll receive a link to the built package in the email address listed in the maintainer field. It usually takes around 30 minutes. As a side effect, win-build also runs R CMD check on the package, so build_win is also useful to check that your package is ok on windows.

Usage

  build_win(pkg = ".", version = "R-release",
    quiet = FALSE)

Arguments

pkg
package description, can be path or package name. See as.package for more information

version
directory to upload to on the win-builder, controlling which version of R is used to build the package. Possible options are listed on http://win-builder.r-project.org/. Defaults to the released version of R.

quiet
if TRUE suppresses output from this function.

See Also

Other build functions: build

[Package devtools version 1.3 Index]