check_cran {devtools}	R Documentation

Check a package from CRAN.

Description

Internal function used to power revdep_check().

Usage

check_cran(pkgs, libpath = file.path(tempdir(), "R-lib"), srcpath = libpath,
  bioconductor = FALSE, type = getOption("pkgType"),
  threads = getOption("Ncpus", 1), check_dir = tempfile("check_cran"),
  revdep_pkg = NULL)

Arguments

pkgs
Vector of package names - note that unlike other devtools functions this is the name of a CRAN package, not a path.

libpath
Path to library to store dependencies packages - if you you're doing this a lot it's a good idea to pick a directory and stick with it so you don't have to download all the packages every time.

srcpath
Path to directory to store source versions of dependent packages - again, this saves a lot of time because you don't need to redownload the packages every time you run the package.

bioconductor
Include bioconductor packages in checking?

type
binary Package type to test (source, mac.binary etc). Defaults to the same type as install.packages().

threads
Number of concurrent threads to use for checking. It defaults to the option "Ncpus" or 1 if unset.

check_dir
Directory to store results.

revdep_pkg
Optional name of a package for which this check is checking the reverse dependencies of. This is normally passed in from revdep_check, and is used only for logging.

Details

This function does not clean up after itself, but does work in a session-specific temporary directory, so all files will be removed when your current R session ends.

Value

invisible TRUE if successful and no ERRORs or WARNINGS.

[Package devtools version 1.7.0 Index]