check_dep_version {devtools}	R Documentation

Check that the version of an imported package satisfies the requirements

Description

Check that the version of an imported package satisfies the requirements

Usage

  check_dep_version(dep_name, dep_ver = NA,
    dep_compare = NA)

Arguments

dep_name
The name of the package with objects to import

dep_ver
The version of the package

dep_compare
The comparison operator to use to check the version

[Package devtools version 1.3 Index]