check_doc {devtools}	R Documentation

Check documentation, as R CMD check does.

Description

This function attempts to run the documentation related checks in the same way that R CMD check does. Unfortunately it can't run them all because some tests require the package to be loaded, and the way they attempt to load the code conflicts with how devtools does it.

Usage

  check_doc(pkg = ".")

Arguments

pkg
package description, can be path or package name. See as.package for more information

Value

Nothing. This function is called purely for it's side effects: if

Examples

## Not run: 
document("mypkg")
check_doc("mypkg")

## End(Not run)

[Package devtools version 1.3 Index]