clean_source {devtools}	R Documentation

Sources an R file in a clean environment.

Description

Opens up a fresh R environment and sources file, ensuring that it works independently of the current working environment.

Usage

  clean_source(path, quiet = FALSE)

Arguments

path
path to R script

quiet
If FALSE, the default, all input and output will be displayed, as if you'd copied and paste the code. If TRUE only the final result and the any explicitly printed output will be displayed.

[Package devtools version 1.3 Index]