cran_env_vars {devtools}	R Documentation

Cran environmental variables. The environment variables that are used CRAN when checking packages. These environment variables are from the R Internals document. The only difference from that document is that here, _R_CHECK_CRAN_INCOMING_ is not set to TRUE because it is so slow.

Description

Cran environmental variables. The environment variables that are used CRAN when checking packages. These environment variables are from the R Internals document. The only difference from that document is that here, _R_CHECK_CRAN_INCOMING_ is not set to TRUE because it is so slow.

Usage

  cran_env_vars()

Value

a named character vector

[Package devtools version 1.3 Index]