create {devtools}	R Documentation

Creates a new package, following all devtools package conventions.

Description

Similar to package.skeleton, except that it only creates the standard devtools directory structures, it doesn't try and create source code and data files by inspecting the global environment.

Usage

  create(path, description = getOption("devtools.desc"),
    check = FALSE)

Arguments

path
location to create new package. The last component of the path will be used as the package name.

description
list of description values to override default values or add additional values.

check
if TRUE, will automatically run check

See Also

Text with package.skeleton

Examples

# Create a package using all defaults:
path <- file.path(tempdir(), "myDefaultPackage")
create(path)

# Override a description attribute.
path <- file.path(tempdir(), "myCustomPackage")
my_description <- list("Maintainer" =
  "'Yoni Ben-Meshulam' <yoni@opower.com>")
create(path, my_description)

[Package devtools version 1.3 Index]