create_description {devtools}	R Documentation

Create a default DESCRIPTION file for a package.

Description

Create a default DESCRIPTION file for a package.

Usage

  create_description(path,
    extra = getOption("devtools.desc"), quiet = FALSE)

Arguments

path
path to package root directory

extra
a named list of extra options to add to ‘DESCRIPTION’

quiet
if TRUE, suppresses output from this function.

Details

To set the default author and licenses, set options devtool.author and devtool.license. I use options(devtools.author = '"Hadley Wickham <h.wickham@gmail.com> [aut,cre]"', devtools.license = "GPL-3").

[Package devtools version 1.3 Index]