dev_help {devtools}	R Documentation

Read the in-development help for a package loaded with devtools.

Description

Note that this only renders a single documentation file, so that links to other files within the package won't work.

Usage

  dev_help(topic, stage = "render",
    type = getOption("help_type"))

Arguments

topic
name of help to search for.

stage
at which stage ("build", "install", or "render") should \Sexpr macros be executed? This is only important if you're using \Sexpr macro's in your Rd files.

type
of html to produce: "html" or "text". Defaults to your default documentation type.

Examples

## Not run: 
library("ggplot2")
help("ggplot") # loads installed documentation for ggplot

load_all("ggplot2")
dev_help("ggplot") # loads development documentation for ggplot

## End(Not run)

[Package devtools version 1.3 Index]