dev_mode {devtools}	R Documentation

Activate and deactivate development mode.

Description

When activated, dev_mode creates a new library for storing installed packages. This new library is automatically created when dev_mode is activated if it does not already exist. This allows you to test development packages in a sandbox, without interfering with the other packages you have installed.

Usage

  dev_mode(on = NULL, path = getOption("devtools.path"))

Arguments

on
turn dev mode on (TRUE) or off (FALSE). If omitted will guess based on whether or not path is in .libPaths

path
directory to library.

Examples

dev_mode()
dev_mode()

[Package devtools version 1.3 Index]