dev_packages {devtools}	R Documentation

Return a vector of names of packages loaded by devtools

Description

Return a vector of names of packages loaded by devtools

Usage

  dev_packages()

[Package devtools version 1.3 Index]