devtools {devtools}	R Documentation

Package development tools for R.

Description

Package development tools for R.

Package options

Devtools uses the following options to configure behaviour:

devtools.path: path to use for dev_mode

devtools.name: your name, used when signing draft emails.

devtools.cleandoc: a single logical value determining whether the contents of the fileman/ directory are deleted by check

devtools.install.args: a string giving extra arguments passed to R CMD install by install.

devtools.desc.author: a string providing a default Authors@R string to be used in new ‘DESCRIPTION’s. Should be a R code, and look like "Hadley Wickham <h.wickham@gmail.com> [aut, cre]". See as.person for more details.

devtools.desc.license: a default license string to use for new packages.

devtools.desc.suggests: a character vector listing packages to to add to suggests by defaults for new packages.

devtools.desc: a named character vector listing any other extra options to add to ‘DESCRIPTION’

[Package devtools version 1.3 Index]