document {devtools}	R Documentation

Use roxygen to make documentation.

Description

Use roxygen to make documentation.

Usage

  document(pkg = ".", clean = FALSE,
    roclets = c("collate", "namespace", "rd"),
    reload = TRUE)

Arguments

pkg
package description, can be path or package name. See as.package for more information

clean
if TRUE will automatically clear all roxygen caches and delete current ‘man/’ contents to ensure that you have the freshest version of the documentation.

roclets
character vector of roclet names to apply to package

reload
if TRUE uses load_all to reload the package prior to documenting. This is important because roxygen2 uses introspection on the code objects to determine how to document them.

[Package devtools version 1.3 Index]