find_rtools {devtools}	R Documentation

Find rtools.

Description

To build binary packages on windows, Rtools (found at http://cran.r-project.org/bin/windows/Rtools/) needs to be on the path. The default installation process does not add it, so this script finds it (looking first on the path, then in the registry). It also checks that the version of rtools matches the version of R.

Usage

  find_rtools(debug = FALSE)

Arguments

debug
if TRUE prints a lot of additional information to help in debugging.

Value

Either a visible TRUE if rtools is found, or an invisible FALSE with a diagnostic message. As a side-effect the internal package variable rtools_path is updated to the paths to rtools binaries.

Acknowledgements

This code borrows heavily from RStudio's code for finding rtools. Thanks JJ!

[Package devtools version 1.3 Index]