git_path {devtools}	R Documentation

Retrieve the current running path of the git binary.

Description

Retrieve the current running path of the git binary.

Usage

  git_path(git_binary_name = NULL)

Arguments

git_binary_name
The name of the binary depending on the OS.

[Package devtools version 1.3 Index]