imports_env {devtools}	R Documentation

Return imports environment for a package

Description

Contains objects imported from other packages. Is the parent of the package namespace environment, and is a child of <namespace:base>, which is a child of R_GlobalEnv.

Usage

  imports_env(pkg = ".")

Arguments

pkg
package description, can be path or package name. See as.package for more information.

See Also

ns_env for the namespace environment that all the objects (exported and not exported).

pkg_env for the attached environment that contains the exported objects.

[Package devtools version 1.3 Index]