infrastructure {devtools}	R Documentation

Add useful infrastructure to a package.

Description

Add useful infrastructure to a package.

Usage

use_testthat(pkg = ".")

use_rstudio(pkg = ".")

use_vignette(name, pkg = ".")

use_rcpp(pkg = ".")

use_travis(pkg = ".")

use_appveyor(pkg = ".")

use_package_doc(pkg = ".")

use_revdep(pkg = ".")

use_cran_comments(pkg = ".")

Arguments

pkg
package description, can be path or package name. See as.package for more information.

name
File name to use for new vignette. Should consist only of numbers, letters, _ and -. I recommend using lower case.

use_testthat

Add testing infrastructure to a package that does not already have it. This will create ‘tests/testthat.R’, ‘tests/testthat/’ and add testthat to the suggested packages. This is called automatically from test if needed.

use_rstudio

Does not modify .Rbuildignore as RStudio will do that when opened for the first time.

use_knitr

Adds needed packages to DESCRIPTION, and creates draft vignette in vignettes/. It adds inst/doc to .gitignore so you don't accidentally check in the built vignettes.

use_rcpp

Creates src/ and adds needed packages to DESCRIPTION.

use_travis

Add basic travis template to a package. Also adds .travis.yml to .Rbuildignore so it isn't included in the built package.

use_appveyor

Add basic AppVeyor template to a package. Also adds appveyor.yml to .Rbuildignore so it isn't included in the built package.

use_package_doc

Adds a roxygen template for package documentation

use_revdep

Add revdep directory and basic check template.

use_cran_comments

Add cran-comments.md template.

See Also

Other infrastructure: add_build_ignore, use_build_ignore; use_data_raw; use_data; use_git_hook; use_package; use_readme_rmd

[Package devtools version 1.7.0 Index]