inst {devtools}	R Documentation

Get the installation path of a package

Description

Given the name of a package, this returns a path to the installed copy of the package, which can be passed to other devtools functions.

Usage

  inst(name)

Arguments

name
the name of a package.

Details

It searches for the package in .libPaths(). If multiple dirs are found, it will return the first one.

Examples

inst("devtools")
inst("grid")
## Not run: 
# Can be passed to other devtools functions
unload(inst("ggplot2"))

## End(Not run)

[Package devtools version 1.3 Index]