Install a package from a Bioconductor repository

This function requires git to be installed on your system in order to be used.

install_bioc(repo, mirror = getOption("BioC_git",
  "https://git.bioconductor.org/packages"), ..., quiet = FALSE)

Arguments

repo
Repository address in the format [username:password@][release/]repo[#commit]. Valid values for the release are ‘devel’ (the default if none specified), ‘release’ or numeric release numbers (e.g. ‘3.3’).

mirror
The bioconductor git mirror to use

...
Other arguments passed on to install

quiet
if TRUE suppresses output from this function.

Details

It is vectorised so you can install multiple packages with a single command.

'

See also

Examples

# NOT RUN {
install_bioc("SummarizedExperiment")
install_bioc("release/SummarizedExperiment")
install_bioc("3.3/SummarizedExperiment")
install_bioc("SummarizedExperiment#abc123")
install_bioc("user:password@release/SummarizedExperiment")
install_bioc("user:password@devel/SummarizedExperiment")
install_bioc("user:password@SummarizedExperiment#abc123")
# }