install_bitbucket {devtools}	R Documentation

Install a package directly from bitbucket

Description

This function is vectorised so you can install multiple packages in a single command.

Usage

install_bitbucket(repo, username, ref = "master", subdir = NULL,
  auth_user = NULL, password = NULL, ...)

Arguments

repo
Repository address in the format username/repo[/subdir][@ref|#pull]. Alternatively, you can specify subdir and/or ref using the respective parameters (see below); if both is specified, the values in repo take precedence.

username
User name. Deprecated: please include username in the repo

ref
Desired git reference; culd be a commit, tag, or branch name. Defaults to master.

subdir
subdirectory within repo that contains the R package.

auth_user
your account username if you're attempting to install a package hosted in a private repository (and your username is different to username)

password
your password

...
Other arguments passed on to install.

See Also

Bitbucket API docs: https://confluence.atlassian.com/display/BITBUCKET/Use+the+Bitbucket+REST+APIs

Other package installation: install_github; install_gitorious; install_git; install_svn; install_url; install_version; install

Examples

## Not run: 
install_bitbucket("sulab/mygene.r@default")
install_bitbucket("dannavarro/lsr-package")

## End(Not run)

[Package devtools version 1.7.0 Index]