install_deps {devtools}	R Documentation

Install package dependencies

Description

Install package dependencies

Usage

install_deps(pkg = ".", dependencies = NA, threads = getOption("Ncpus",
  1))

Arguments

pkg
package description, can be path or package name. See as.package for more information

dependencies
logical indicating to also install uninstalled packages which this pkg depends on/links to/suggests. See argument dependencies of install.packages.

threads
number of concurrent threads to use for installing dependencies. It defaults to the option "Ncpus" or 1 if unset.

Examples

## Not run: install_deps(".")

[Package devtools version 1.7.0 Index]