install_git {devtools}	R Documentation

Install a package from a git repository

Description

It is vectorised so you can install multiple packages with a single command.

Usage

install_git(url, subdir = NULL, branch = NULL, args = character(0), ...)

Arguments

url
Location of package. The url should point to a public or private repository.

subdir
A sub-directory within a git repository that may contain the package we are interested in installing.

branch
Name of branch or tag to use, if not master.

args
A character vector providing extra arguments to pass on to

...
passed on to install

See Also

Other package installation: install_bitbucket; install_github; install_gitorious; install_svn; install_url; install_version; install

Examples

## Not run: 
install_git("git://github.com/hadley/stringr.git")
install_git("git://github.com/hadley/stringr.git", branch = "stringr-0.2")

## End(Not run)

[Package devtools version 1.7.0 Index]