install_git_single {devtools}	R Documentation

Install a single package from a git repository

Description

This function allows you to install a single package from a git repository.

Usage

  install_git_single(git_url, name = NULL, subdir = NULL,
    branch = NULL, git_args = character(),
    git_binary = NULL, ...)

Arguments

git_url
Location of package. The url should point to a public or private repository.

name
Optional package name, used to provide more informative messages.

subdir
A sub-directory withing a git repository that may contain the package we are interested in installing.

git_args
A character vector providing extra arguments to pass on to

git_binary
A custom git-binary to use instead of default system's git version.

...
passed on to install

Details

See install_git for more information about the paraemeters.

[Package devtools version 1.3 Index]