install_gitorious {devtools}	R Documentation

Attempts to install a package directly from gitorious.

Description

This function is vectorised so you can install multiple packages in a single command.

Usage

install_gitorious(repo, ref = "master", subdir = NULL, ...)

Arguments

repo
Repository address in the format username/repo[/subdir][@ref|#pull]. Alternatively, you can specify subdir and/or ref using the respective parameters (see below); if both is specified, the values in repo take precedence.

ref
Desired git reference. Could be a commit, tag, or branch name, or a call to github_pull. Defaults to "master".

subdir
subdirectory within repo that contains the R package.

...
Other arguments passed on to install.

See Also

Other package installation: install_bitbucket; install_github; install_git; install_svn; install_url; install_version; install

Examples

## Not run: 
install_gitorious("r-mpc-package/r-mpc-package")

## End(Not run)

[Package devtools version 1.7.0 Index]