install_local {devtools}	R Documentation

Install a package from a local file

Description

This function is vectorised so you can install multiple packages in a single command.

Usage

  install_local(path, subdir = NULL, ...)

Arguments

path
path to local directory, or compressed file (tar, zip, tar.gz tar.bz2, tgz2 or tbz)

subdir
subdirectory within url bundle that contains the R package.

...
Other arguments passed on to install.

Examples

## Not run: 
dir <- tempfile()
dir.create(dir)
pkg <- download.packages("testthat", dir)
install_local(pkg[, 2])

## End(Not run)

[Package devtools version 1.3 Index]