install_url {devtools}	R Documentation

Install a package from a url

Description

This function is vectorised so you can install multiple packages in a single command.

Usage

  install_url(url, name = NULL, subdir = NULL,
    config = list(), ...)

Arguments

url
location of package on internet. The url should point to a zip file, a tar file or a bzipped/gzipped tar file.

name
optional package name, used to provide more informative messages

subdir
subdirectory within url bundle that contains the R package.

config
additional configuration argument (e.g. proxy, authentication) passed on to GET.

...
Other arguments passed on to install.

See Also

Other package installation: install, install_bitbucket, install_git, install_github, install_gitorious, install_version

[Package devtools version 1.3 Index]