install_version {devtools}	R Documentation

Install specified version of a CRAN package.

Description

If you are installing an package that contains compiled code, you will need to have an R development environment installed. You can check if you do by running has_devel.

Usage

install_version(package, version = NULL, repos = getOption("repos"),
  type = getOption("pkgType"), ...)

Arguments

package
package name

version
If the specified version is NULL or the same as the most recent version of the package, this function simply calls install. Otherwise, it looks at the list of archived source tarballs and tries to install an older version instead.

repos
character vector, the base URL(s) of the repositories to use, e.g., the URL of a CRAN mirror such as "http://cran.us.r-project.org". For more details on supported URL schemes see url.

Can be NULL to install from local files, directories or URLs: this will be inferred by extension from pkgs if of length one.

type
character, indicating the type of package to download and install.

Possible values are (currently) "source", "mac.binary", "mac.binary.mavericks" and "win.binary": the binary types can be listed and downloaded but not installed on other platforms.

The default is the appropriate binary type on Windows and on the CRAN binary OS X distributions, otherwise "source". For the platforms where binary packages are the default, an alternative is "both" which means ‘try binary if available and current, otherwise try source’. (This will only choose the binary package if its version number is no older than the source version. In interactive use it will ask before attempting to install source packages.)

...
Other arguments passed on to install.

Author(s)

Jeremy Stephens

See Also

Other package installation: install_bitbucket; install_github; install_gitorious; install_git; install_svn; install_url; install

[Package devtools version 1.7.0 Index]