lint {devtools}	R Documentation

Lint all source files in a package.

Description

The default lintings correspond to the style guide at http://r-pkgs.had.co.nz/r.html#style, however it is possible to override any or all of them using the linters parameter.

Usage

lint(pkg = ".", ...)

Arguments

pkg
package description, can be path or package name. See as.package for more information

...
additional arguments passed to lint_package

See Also

lint_package, lint

[Package devtools version 1.7.0 Index]