loaded_packages {devtools}	R Documentation

Return a vector of names of attached packages

Description

Return a vector of names of attached packages

Usage

loaded_packages()

Value

A data frame with columns package and path, giving the name of each package and the path it was loaded from.

[Package devtools version 2.4.2 Index]