ns_env {devtools}	R Documentation

Return the namespace environment for a package.

Description

Contains all (exported and non-exported) objects, and is a descendent of R_GlobalEnv. The hieararchy is <namespace:pkg>, <imports:pkg>, <namespace:base>, and then R_GlobalEnv.

Usage

  ns_env(pkg = ".")

Arguments

pkg
package description, can be path or package name. See as.package for more information

Details

If the package is not loaded, this function returns NULL.

See Also

pkg_env for the attached environment that contains the exported objects.

imports_env for the environment that contains imported objects for the package.

[Package devtools version 1.3 Index]