r_env_vars {devtools}	R Documentation

Environment variables to set when calling R

Description

Devtools sets a number of environmental variables to ensure consistent between the current R session and the new session, and to ensure that everying behaves the same across systems. It also suppresses a common warning on windows, and sets NOT_CRAN so you can tell that your code is not running on CRAN.

Usage

  r_env_vars()

Value

a named character vector

[Package devtools version 1.3 Index]