release {devtools}	R Documentation

Release package to CRAN.

Description

Run automated and manual tests, then ftp to CRAN.

Usage

  release(pkg = ".", check = TRUE)

Arguments

pkg
package description, can be path or package name. See as.package for more information

check
if TRUE, run checking, otherwise omit it. This is useful if you've just checked your package and you're ready to release it.

Details

The package release process will:

Confirm that the package passes R CMD check

Ask if you've checked your code on win-builder

Confirm that news is up-to-date

Confirm that DESCRIPTION is ok

Ask if you've checked packages that depend on your package

Build the package

Upload the package to CRAN

Draft an email to the CRAN maintainer.

You also need to read the CRAN repository policy at http://cran.r-project.org/web/packages/policies.html and make sure you're in line with the policies. release tries to automate as many of polices as possible, but it's impossible to be completely comprehensive, and they do change in between releases of devtools.

Guarantee

If a devtools bug causes one of the CRAN maintainers to treat you impolitely, I will personally send you a handwritten apology note. Please forward me the email and your address, and I'll get a card in the mail.

[Package devtools version 1.3 Index]