reload {devtools}	R Documentation

Unload and reload package.

Description

This attempts to unload and reload a package. If the package is not loaded already, it does nothing. It's not always possible to cleanly unload a package: see the caveats in unload for the some of the potential failure points. If in doubt, restart R and reload the package with library.

Usage

  reload(pkg = ".", quiet = FALSE)

Arguments

pkg
package description, can be path or package name. See as.package for more information

quiet
if TRUE suppresses output from this function.

Examples

## Not run: 
# Reload package that is in current directory
reload(".")

# Reload package that is in ./ggplot2/
reload("ggplot2/")

# Can use inst() to find the package path
# This will reload the installed ggplot2 package
reload(inst("ggplot2"))

## End(Not run)

[Package devtools version 1.3 Index]