revdep {devtools}	R Documentation

Reverese dependency tools.

Description

Tools to check and notify maintainers of all all CRAN and bioconductor packages that depend on the specified package.

Usage

  revdep(pkg = NULL,
    dependencies = c("Depends", "Imports", "Suggests", "LinkingTo"),
    recursive = FALSE)

  revdep_maintainers(pkg = ".")

  revdep_check(pkg = NULL, recursive = FALSE, ...)

Arguments

pkg
package name

...
Other parameters passed on to check_cran

dependencies
a character vector listing the types of dependencies, a subset of c("Depends", "Imports", "LinkingTo", "Suggests", "Enhances"). Character string "all" is shorthand for that vector.

recursive
logical: should reverse dependencies of reverse dependencies (and so on) be included?

Examples

## Not run: revdep("ggplot2")

[Package devtools version 1.3 Index]