Print session information

This is sessionInfo() re-written from scratch to both exclude data that's rarely useful (e.g., the full collate string or base packages loaded) and include stuff you'd like to know (e.g., where a package was installed from).

session_info(pkgs = NULL, include_base = FALSE)

Arguments

pkgs
Either a vector of package names or NULL. If NULL, displays all loaded packages. If a character vector, also, includes all dependencies of the package.

include_base
Include base packages in summary? By default this is false since base packages should always match the R version.

Examples

session_info()

#> Session info ------------------------------------------------------------------

#>  setting  value                       
#>  version  R version 3.5.0 (2018-04-23)
#>  os       macOS High Sierra 10.13.4   
#>  system   x86_64, darwin17.5.0        
#>  ui       RStudio                     
#>  language (EN)                        
#>  collate  en_US.UTF-8                 
#>  tz       America/Denver              
#>  date     2018-05-14                  
#> 

#> Packages ----------------------------------------------------------------------

#>  package     ondiskversion loadedversion
#>  assertthat  0.2.0         0.2.0        
#>  backports   1.1.2         1.1.2        
#>  callr       2.0.3         2.0.3        
#>  cli         1.0.0         1.0.0        
#>  clisymbols  1.2.0         1.2.0        
#>  commonmark  1.5           1.5          
#>  crayon      1.3.4         1.3.4        
#>  curl        3.2           3.2          
#>  debugme     1.1.0         1.1.0        
#>  desc        1.2.0         1.2.0        
#>  devtools    1.13.5.9000   1.13.5.9000  
#>  digest      0.6.15        0.6.15       
#>  evaluate    0.10.1        0.10.1       
#>  fs          1.2.2         1.2.2        
#>  git2r       0.21.0.9002   0.21.0       
#>  highlight   0.4.7.2       0.4.7.2      
#>  htmltools   0.3.6         0.3.6        
#>  httr        1.3.1         1.3.1        
#>  jsonlite    1.5           1.5          
#>  knitr       1.20          1.20         
#>  magrittr    1.5           1.5          
#>  MASS        7.3.50        7.3-50       
#>  memoise     1.1.0         1.1.0        
#>  mime        0.5           0.5          
#>  openssl     1.0.1         1.0.1        
#>  pillar      1.2.2         1.2.2        
#>  pkgbuild    1.0.0         1.0.0        
#>  pkgdown     1.0.0.9000    1.0.0.9000   
#>  pkgload     1.0.0         1.0.0        
#>  prettyunits 1.0.2         1.0.2        
#>  purrr       0.2.4         0.2.4        
#>  R6          2.2.2         2.2.2        
#>  rcmdcheck   1.2.1.9005    1.2.1.9005   
#>  Rcpp        0.12.16       0.12.16      
#>  rematch     1.0.1         1.0.1        
#>  remotes     1.1.1         1.1.1        
#>  rlang       0.2.0         0.2.0        
#>  rmarkdown   1.9           1.9          
#>  roxygen2    6.0.1         6.0.1        
#>  rprojroot   1.3.2         1.3-2        
#>  rstudioapi  0.7           0.7          
#>  sessioninfo 1.0.0         1.0.0        
#>  stringi     1.2.2         1.2.2        
#>  stringr     1.3.0         1.3.0        
#>  testthat    2.0.0         2.0.0        
#>  tibble      1.4.2         1.4.2        
#>  usethis     1.3.0         1.3.0        
#>  whisker     0.3.2         0.3-2        
#>  withr       2.1.2         2.1.2        
#>  xml2        1.2.0         1.2.0        
#>  yaml        2.1.19        2.1.19       
#>  path                                           attached is_base date      
#>  /usr/local/lib/R/3.5/site-library/assertthat   FALSE    FALSE   2017-04-11
#>  /usr/local/lib/R/3.5/site-library/backports    FALSE    FALSE   2017-12-13
#>  /usr/local/lib/R/3.5/site-library/callr        FALSE    FALSE   2018-04-11
#>  /usr/local/lib/R/3.5/site-library/cli          FALSE    FALSE   2017-11-05
#>  /usr/local/lib/R/3.5/site-library/clisymbols   FALSE    FALSE   2017-05-21
#>  /usr/local/lib/R/3.5/site-library/commonmark   FALSE    FALSE   2018-04-28
#>  /usr/local/lib/R/3.5/site-library/crayon       FALSE    FALSE   2017-09-16
#>  /usr/local/lib/R/3.5/site-library/curl         FALSE    FALSE   2018-03-28
#>  /usr/local/lib/R/3.5/site-library/debugme      FALSE    FALSE   2017-10-22
#>  /usr/local/lib/R/3.5/site-library/desc         FALSE    FALSE   2018-05-01
#>  /Users/jayhesselberth/devel/devtools            TRUE    FALSE   <NA>      
#>  /usr/local/lib/R/3.5/site-library/digest       FALSE    FALSE   2018-01-28
#>  /usr/local/lib/R/3.5/site-library/evaluate     FALSE    FALSE   2017-06-24
#>  /usr/local/lib/R/3.5/site-library/fs           FALSE    FALSE   2018-03-21
#>  /usr/local/lib/R/3.5/site-library/git2r        FALSE    FALSE   2018-05-14
#>  /usr/local/lib/R/3.5/site-library/highlight    FALSE    FALSE   2017-10-04
#>  /usr/local/lib/R/3.5/site-library/htmltools    FALSE    FALSE   2017-04-28
#>  /usr/local/lib/R/3.5/site-library/httr         FALSE    FALSE   2017-08-20
#>  /usr/local/lib/R/3.5/site-library/jsonlite     FALSE    FALSE   2017-06-01
#>  /usr/local/lib/R/3.5/site-library/knitr        FALSE    FALSE   2018-02-20
#>  /usr/local/lib/R/3.5/site-library/magrittr     FALSE    FALSE   2014-11-22
#>  /usr/local/Cellar/r/3.5.0_1/lib/R/library/MASS FALSE    FALSE   2018-04-30
#>  /usr/local/lib/R/3.5/site-library/memoise      FALSE    FALSE   2017-04-21
#>  /usr/local/lib/R/3.5/site-library/mime         FALSE    FALSE   2016-07-07
#>  /usr/local/lib/R/3.5/site-library/openssl      FALSE    FALSE   2018-03-03
#>  /usr/local/lib/R/3.5/site-library/pillar       FALSE    FALSE   2018-04-26
#>  /usr/local/lib/R/3.5/site-library/pkgbuild     FALSE    FALSE   2018-05-14
#>  /usr/local/lib/R/3.5/site-library/pkgdown      FALSE    FALSE   2018-05-11
#>  /usr/local/lib/R/3.5/site-library/pkgload      FALSE    FALSE   2018-05-14
#>  /usr/local/lib/R/3.5/site-library/prettyunits  FALSE    FALSE   2015-07-13
#>  /usr/local/lib/R/3.5/site-library/purrr        FALSE    FALSE   2017-10-18
#>  /usr/local/lib/R/3.5/site-library/R6           FALSE    FALSE   2017-06-17
#>  /usr/local/lib/R/3.5/site-library/rcmdcheck    FALSE    FALSE   2018-05-14
#>  /usr/local/lib/R/3.5/site-library/Rcpp         FALSE    FALSE   2018-03-13
#>  /usr/local/lib/R/3.5/site-library/rematch      FALSE    FALSE   2016-04-21
#>  /usr/local/lib/R/3.5/site-library/remotes      FALSE    FALSE   2017-12-20
#>  /usr/local/lib/R/3.5/site-library/rlang        FALSE    FALSE   2018-02-20
#>  /usr/local/lib/R/3.5/site-library/rmarkdown    FALSE    FALSE   2018-03-01
#>  /usr/local/lib/R/3.5/site-library/roxygen2     FALSE    FALSE   2017-02-06
#>  /usr/local/lib/R/3.5/site-library/rprojroot    FALSE    FALSE   2018-01-03
#>  /usr/local/lib/R/3.5/site-library/rstudioapi   FALSE    FALSE   2017-09-07
#>  /usr/local/lib/R/3.5/site-library/sessioninfo  FALSE    FALSE   2017-06-21
#>  /usr/local/lib/R/3.5/site-library/stringi      FALSE    FALSE   2018-05-02
#>  /usr/local/lib/R/3.5/site-library/stringr      FALSE    FALSE   2018-02-19
#>  /usr/local/lib/R/3.5/site-library/testthat      TRUE    FALSE   2017-12-13
#>  /usr/local/lib/R/3.5/site-library/tibble       FALSE    FALSE   2018-01-22
#>  /usr/local/lib/R/3.5/site-library/usethis       TRUE    FALSE   2018-02-24
#>  /usr/local/lib/R/3.5/site-library/whisker      FALSE    FALSE   2013-04-28
#>  /usr/local/lib/R/3.5/site-library/withr        FALSE    FALSE   2018-03-15
#>  /usr/local/lib/R/3.5/site-library/xml2         FALSE    FALSE   2018-01-24
#>  /usr/local/lib/R/3.5/site-library/yaml         FALSE    FALSE   2018-05-01
#>  source                          
#>  CRAN (R 3.5.0)                  
#>  CRAN (R 3.5.0)                  
#>  CRAN (R 3.5.0)                  
#>  CRAN (R 3.5.0)                  
#>  CRAN (R 3.5.0)                  
#>  CRAN (R 3.5.0)                  
#>  CRAN (R 3.5.0)                  
#>  CRAN (R 3.5.0)                  
#>  CRAN (R 3.5.0)                  
#>  CRAN (R 3.5.0)                  
#>  local                           
#>  CRAN (R 3.5.0)                  
#>  CRAN (R 3.5.0)                  
#>  CRAN (R 3.5.0)                  
#>  Github (ropensci/git2r@0eaa7fb) 
#>  CRAN (R 3.5.0)                  
#>  CRAN (R 3.5.0)                  
#>  CRAN (R 3.5.0)                  
#>  CRAN (R 3.5.0)                  
#>  CRAN (R 3.5.0)                  
#>  CRAN (R 3.5.0)                  
#>  CRAN (R 3.5.0)                  
#>  CRAN (R 3.5.0)                  
#>  CRAN (R 3.5.0)                  
#>  CRAN (R 3.5.0)                  
#>  CRAN (R 3.5.0)                  
#>  Github (r-lib/pkgbuild@0457039) 
#>  local                           
#>  Github (r-lib/pkgload@35efedd)  
#>  CRAN (R 3.5.0)                  
#>  CRAN (R 3.5.0)                  
#>  CRAN (R 3.5.0)                  
#>  Github (r-lib/rcmdcheck@ad42561)
#>  CRAN (R 3.5.0)                  
#>  CRAN (R 3.5.0)                  
#>  CRAN (R 3.5.0)                  
#>  CRAN (R 3.5.0)                  
#>  CRAN (R 3.5.0)                  
#>  CRAN (R 3.5.0)                  
#>  CRAN (R 3.5.0)                  
#>  CRAN (R 3.5.0)                  
#>  CRAN (R 3.5.0)                  
#>  CRAN (R 3.5.0)                  
#>  CRAN (R 3.5.0)                  
#>  CRAN (R 3.5.0)                  
#>  CRAN (R 3.5.0)                  
#>  CRAN (R 3.5.0)                  
#>  CRAN (R 3.5.0)                  
#>  CRAN (R 3.5.0)                  
#>  CRAN (R 3.5.0)                  
#>  CRAN (R 3.5.0)                  

session_info("devtools")

#> Session info ------------------------------------------------------------------

#>  setting  value                       
#>  version  R version 3.5.0 (2018-04-23)
#>  os       macOS High Sierra 10.13.4   
#>  system   x86_64, darwin17.5.0        
#>  ui       RStudio                     
#>  language (EN)                        
#>  collate  en_US.UTF-8                 
#>  tz       America/Denver              
#>  date     2018-05-14                  
#> 

#> Packages ----------------------------------------------------------------------

#>  package    ondiskversion loadedversion
#>  curl       3.2           3.2          
#>  devtools   1.13.5.9000   1.13.5.9000  
#>  digest     0.6.15        0.6.15       
#>  git2r      0.21.0.9002   0.21.0       
#>  httr       1.3.1         1.3.1        
#>  jsonlite   1.5           1.5          
#>  memoise    1.1.0         1.1.0        
#>  mime       0.5           0.5          
#>  openssl    1.0.1         1.0.1        
#>  R6         2.2.2         2.2.2        
#>  rstudioapi 0.7           0.7          
#>  whisker    0.3-2         0.3-2        
#>  withr      2.1.2         2.1.2        
#>  path                                         attached is_base date      
#>  /usr/local/lib/R/3.5/site-library/curl       FALSE    FALSE   2018-03-28
#>  /Users/jayhesselberth/devel/devtools          TRUE    FALSE   <NA>      
#>  /usr/local/lib/R/3.5/site-library/digest     FALSE    FALSE   2018-01-28
#>  /usr/local/lib/R/3.5/site-library/git2r      FALSE    FALSE   2018-05-14
#>  /usr/local/lib/R/3.5/site-library/httr       FALSE    FALSE   2017-08-20
#>  /usr/local/lib/R/3.5/site-library/jsonlite   FALSE    FALSE   2017-06-01
#>  /usr/local/lib/R/3.5/site-library/memoise    FALSE    FALSE   2017-04-21
#>  /usr/local/lib/R/3.5/site-library/mime       FALSE    FALSE   2016-07-07
#>  /usr/local/lib/R/3.5/site-library/openssl    FALSE    FALSE   2018-03-03
#>  /usr/local/lib/R/3.5/site-library/R6         FALSE    FALSE   2017-06-17
#>  /usr/local/lib/R/3.5/site-library/rstudioapi FALSE    FALSE   2017-09-07
#>  /usr/local/lib/R/3.5/site-library/whisker    FALSE    FALSE   2013-04-28
#>  /usr/local/lib/R/3.5/site-library/withr      FALSE    FALSE   2018-03-15
#>  source                         
#>  CRAN (R 3.5.0)                 
#>  local                          
#>  CRAN (R 3.5.0)                 
#>  Github (ropensci/git2r@0eaa7fb)
#>  CRAN (R 3.5.0)                 
#>  CRAN (R 3.5.0)                 
#>  CRAN (R 3.5.0)                 
#>  CRAN (R 3.5.0)                 
#>  CRAN (R 3.5.0)                 
#>  CRAN (R 3.5.0)                 
#>  CRAN (R 3.5.0)                 
#>  CRAN (R 3.5.0)                 
#>  CRAN (R 3.5.0)