spell_check {devtools}	R Documentation

Spell checking

Description

Runs a spell check on text fields in the package description file, manual pages, and optionally vignettes. Wraps the spelling package.

Usage

spell_check(pkg = ".", vignettes = TRUE, use_wordlist = TRUE)

Arguments

pkg
The package to use, can be a file path to the package or a package object. See as.package() for more information.

vignettes
also check all rmd and rnw files in the pkg vignettes folder

use_wordlist
ignore words in the package WORDLIST file

[Package devtools version 2.4.2 Index]