submit_cran {devtools}	R Documentation

Submit a package to CRAN.

Description

This uses the new CRAN web-form submission process. After submission, you will receive an email asking you to confirm submission - this is used to check that the package is submitted by the maintainer.

Usage

submit_cran(pkg = ".")

Arguments

pkg
package description, can be path or package name. See as.package for more information

Details

It's recommend that you use release() rather than this function as it performs more checks prior to submission.

[Package devtools version 1.7.0 Index]