system_check {devtools}	R Documentation

Run a system command and check if it succeeds.

Description

Run a system command and check if it succeeds.

Usage

system_check(cmd, args = character(), env_vars = character(),
  path = ".", quiet = FALSE, throw = TRUE, ...)

Arguments

cmd
Command to run. Will be quoted by shQuote().

args
A character vector of arguments.

env_vars
A named character vector of environment variables.

path
Path in which to execute the command

quiet
If FALSE, the command to be run will be echoed.

throw
If TRUE, will throw an error if the command fails (i.e. the return value is not 0).

...
additional arguments passed to base::system()

Value

The exit status of the command, invisibly.

[Package devtools version 2.0.2 Index]