system_output {devtools}	R Documentation

Run a system command and capture the output.

Description

Run a system command and capture the output.

Usage

system_output(cmd, args = character(), env_vars = character(),
  path = ".", quiet = FALSE, ...)

Arguments

cmd
Command to run. Will be quoted by shQuote().

args
A character vector of arguments.

env_vars
A named character vector of environment variables.

path
Path in which to execute the command

quiet
If FALSE, the command to be run will be echoed.

...
additional arguments passed to base::system()

Value

command output if the command succeeds, an error will be thrown if the command fails.

[Package devtools version 2.0.2 Index]