test {devtools}	R Documentation

Execute all test_that tests in a package.

Description

Tests are assumed to be located in a inst/tests/ directory. See test_dir for the naming convention of test scripts within that directory.

Usage

  test(pkg = ".", filter = NULL, fresh = FALSE)

Arguments

pkg
package description, can be path or package name. See as.package for more information

filter
If not NULL, only tests with file names matching this regular expression will be executed. Matching will take on the file name after it has been stripped of "test-" and ".r".

fresh
if TRUE, will be run in a fresh R session. This has the advantage that there's no way the examples can depend on anything in the current session, but interactive code (like browser) won't work.

[Package devtools version 1.3 Index]