uninstall {devtools}	R Documentation

Uninstall a local development package.

Description

Uses remove.package to uninstall the package. To uninstall a package from a non-default library, use withr::with_libpaths().

Usage

uninstall(pkg = ".", unload = TRUE, quiet = FALSE, lib = .libPaths()[[1]])

Arguments

pkg
The package to use, can be a file path to the package or a package object. See as.package() for more information.

unload
if TRUE (the default), will automatically unload the package prior to uninstalling.

quiet
If TRUE, suppress output.

lib
a character vector giving the library directories to remove the packages from. If missing, defaults to the first element in .libPaths().

See Also

with_debug() to install packages with debugging flags set.

Other package installation: install()

[Package devtools version 2.4.2 Index]