Update packages that are missing or out-of-date.

Works similarly to install.packages() but doesn't install packages that are already installed, and also upgrades out dated dependencies.

update_packages(pkgs = NULL, dependencies = NA,
  repos = getOption("repos"), type = getOption("pkgType"))

Arguments

pkgs
Character vector of packages to update. IF TRUE all installed packages are updated. If NULL user is prompted to confirm update of all installed packages.

dependencies
Which dependencies do you want to check? Can be a character vector (selecting from "Depends", "Imports", "LinkingTo", "Suggests", or "Enhances"), or a logical vector.

TRUE is shorthand for "Depends", "Imports", "LinkingTo" and "Suggests". NA is shorthand for "Depends", "Imports" and "LinkingTo" and is the default. FALSE is shorthand for no dependencies (i.e. just check this package, not its dependencies).

repos
A character vector giving repositories to use.

type
Type of package to update.

See also

Examples

# NOT RUN {
update_packages("ggplot2")
update_packages(c("plyr", "ggplot2"))
# }