use_build_ignore {devtools}	R Documentation

Add a file to .Rbuildignore

Description

.Rbuildignore has a regular expression on each line, but it's usually easier to work with specific file names. By default, will (crudely) turn a filename into a regular expression that will only match that path. Repeated entries will be silently removed.

Usage

use_build_ignore(files, escape = TRUE, pkg = ".")

Arguments

files
Name of file.

escape
If TRUE, the default, will escape . to \. and surround with ^ and $.

pkg
package description, can be path or package name. See as.package for more information

Value

Nothing, called for its side effect.

See Also

Other infrastructure: add_rstudio_project, add_test_infrastructure, add_travis, add_travis, add_travis, infrastructure, use_appveyor, use_cran_comments, use_package_doc, use_rcpp, use_revdep, use_rstudio, use_testthat, use_travis, use_vignette; use_data_raw; use_data; use_git_hook; use_package; use_readme_rmd

[Package devtools version 1.7.0 Index]