Create package pdf manual

Create package pdf manual

Usage

build_manual(pkg = ".", path = NULL)

Arguments

pkg

The package to use, can be a file path to the package or a package object. See as.package() for more information.

path

path in which to produce package manual. If NULL, defaults to the parent directory of the package.

See also