check_cran {devtools}	R Documentation

Check a package from CRAN.

Description

This is useful for automatically checking that dependencies of your packages work.

Usage

  check_cran(pkgs, libpath = file.path(tempdir(), "R-lib"),
    srcpath = libpath, bioconductor = FALSE,
    type = getOption("pkgType"), threads = 1, ...)

Arguments

pkgs
Vector of package names - note that unlike other devtools functions this is the name of a CRAN package, not a path.

libpath
path to library to store dependencies packages - if you you're doing this a lot it's a good idea to pick a directory and stick with it so you don't have to download all the packages every time.

srcpath
path to directory to store source versions of dependent packages - again, this saves a lot of time because you don't need to redownload the packages every time you run the package.

bioconductor
include bioconductor packages in checking?

type
binary package type of test

threads
number of concurrent threads to use for checking.

...
other parameters passed onto download.packages

Details

This function does not clean up after itself, but does work in a session-specific temporary directory, so all files will be removed when your current R session ends.

Value

invisible TRUE if successful and no ERRORs or WARNINGS,

Examples

## Not run: 
dep <- revdep("ggplot2")
check_cran(dep, "~/documents/ggplot/ggplot-check")
# Or, equivalently:
revdep_check("ggplot2")

## End(Not run)

[Package devtools version 1.3 Index]