check_mac_release {devtools}	R Documentation

Check macOS package

Description

This function works by bundling source package, and then uploading to https://mac.r-project.org/macbuilder/submit.html. This function returns a link to the page with the check results.

Usage

check_mac_release(
  pkg = ".",
  dep_pkgs = ".",
  args = NULL,
  manual = TRUE,
  quiet = FALSE,
  ...
)

Arguments

pkg
The package to use, can be a file path to the package or a package object. See as.package() for more information.

dep_pkgs
Additional custom dependencies to install prior to checking the package.

args
An optional character vector of additional command line arguments to be passed to R CMD build if binary = FALSE, or R CMD install if binary = TRUE.

manual
For source packages: if FALSE, don't build PDF vignettes (--no-build-vignettes) or manual (--no-manual).

quiet
If TRUE, suppresses output.

...
Additional arguments passed to pkgbuild::build().

Value

The url with the check results (invisibly)

See Also

Other build functions: check_rhub(), check_win()

[Package devtools version 2.4.3 Index]