check_win {devtools}	R Documentation

Build windows binary package.

Description

This function works by bundling source package, and then uploading to https://win-builder.r-project.org/. Once building is complete you'll receive a link to the built package in the email address listed in the maintainer field. It usually takes around 30 minutes. As a side effect, win-build also runs R CMD check on the package, so check_win is also useful to check that your package is ok on windows.

Usage

check_win_devel(
  pkg = ".",
  args = NULL,
  manual = TRUE,
  email = NULL,
  quiet = FALSE,
  ...
)

check_win_release(
  pkg = ".",
  args = NULL,
  manual = TRUE,
  email = NULL,
  quiet = FALSE,
  ...
)

check_win_oldrelease(
  pkg = ".",
  args = NULL,
  manual = TRUE,
  email = NULL,
  quiet = FALSE,
  ...
)

Arguments

pkg
The package to use, can be a file path to the package or a package object. See as.package() for more information.

args
An optional character vector of additional command line arguments to be passed to R CMD build if binary = FALSE, or R CMD install if binary = TRUE.

manual
For source packages: if FALSE, don't build PDF vignettes (--no-build-vignettes) or manual (--no-manual).

email
An alternative email to use, default NULL uses the package Maintainer's email.

quiet
If TRUE, suppresses output.

...
Additional arguments passed to pkgbuild::build().

Functions

check_win_devel: Check package on the development version of R.

check_win_release: Check package on the release version of R.

check_win_oldrelease: Check package on the previous major release version of R.

See Also

Other build functions: check_rhub()

[Package devtools version 2.4.2 Index]