create {devtools}	R Documentation

Create a package

Description

Create a package

Usage

create(path, ..., open = FALSE)

Arguments

path
A path. If it exists, it is used. If it does not exist, it is created, provided that the parent path exists.

...
Additional arguments passed to usethis::create_package()

open
If TRUE, activates the new project:

If RStudio desktop, the package is opened in a new session.

If on RStudio server, the current RStudio project is activated.

Otherwise, the working directory and active project is changed.

Value

The path to the created package, invisibly.

[Package devtools version 2.4.2 Index]