Create a package

Create a package

create(path, fields = NULL, rstudio = rstudioapi::isAvailable(),
  open = interactive())

Arguments

path
A path. If it exists, it is used. If it does not exist, it is created, provided that the parent path exists.

fields
A named list of fields to add to DESCRIPTION, potentially overriding default values. See use_description() for how you can set personalized defaults using package options

rstudio
If TRUE, calls use_rstudio() to make the new package or into an RStudio Project.

open
If TRUE, activates the new project:

If RStudio desktop, the package is opened in a new session.

If on RStudio server, the current RStudio project is activated.

Otherwise, the working directory and active project is changed.

Value

The path to the created package, invisibly.