dev_packages {devtools}	R Documentation

Return a vector of names of packages loaded by devtools

Description

Return a vector of names of packages loaded by devtools

Usage

dev_packages()

[Package devtools version 2.4.2 Index]