Package development tools for R.

Collection of package development tools.

Package options

Devtools uses the following options() to configure behaviour:

devtools.path: path to use for dev_mode()

devtools.name: your name, used when signing draft emails.

devtools.install.args: a string giving extra arguments passed to R CMD install by install().

devtools.desc.author: a string providing a default Authors@R string to be used in new DESCRIPTIONs. Should be a R code, and look like "Hadley Wickham <h.wickham@gmail.com> [aut, cre]". See utils::as.person() for more details.

devtools.desc.license: a default license string to use for new packages.

devtools.desc.suggests: a character vector listing packages to to add to suggests by defaults for new packages.

devtools.desc: a named list listing any other extra options to add to DESCRIPTION

See also

Author

Maintainer: Jennifer Bryan jenny@rstudio.com (ORCID)

Authors:

Hadley Wickham

Jim Hester

Winston Chang

Other contributors:

RStudio [copyright holder, funder]