test_file {devtools}	R Documentation

Deprecated Functions

Description

These functions are Deprecated in this release of devtools, they will be marked as Defunct and removed in a future version.

Usage

test_file(file = find_active_file(), ...)

test_coverage_file(file = find_active_file(), ...)

[Package devtools version 2.4.2 Index]