document {devtools}	R Documentation

Use roxygen to document a package.

Description

This function is a wrapper for the roxygen2::roxygenize() function from the roxygen2 package. See the documentation and vignettes of that package to learn how to use roxygen.

Usage

document(pkg = ".", roclets = NULL)

Arguments

pkg
package description, can be path or package name. See as.package() for more information

roclets
Character vector of roclet names to use with package. This defaults to NULL, which will use the roclets fields in the list provided in the Roxygen DESCRIPTION field. If none are specified, defaults to c("collate", "namespace", "rd").

See Also

roxygen2::roxygenize(), browseVignettes("roxygen2")

[Package devtools version 2.0.2 Index]