git_checks {devtools}	R Documentation

Git checks.

Description

This function performs Git checks checks prior to release. It is called automatically by release().

Usage

git_checks(pkg = ".")

Arguments

pkg
The package to use, can be a file path to the package or a package object. See as.package() for more information.

[Package devtools version 2.4.2 Index]