install {devtools}	R Documentation

Install a local development package.

Description

Uses R CMD INSTALL to install the package. Will also try to install dependencies of the package from CRAN, if they're not already installed.

Usage

  install(pkg = ".", reload = TRUE, quick = FALSE,
    local = TRUE,
    args = getOption("devtools.install.args"),
    quiet = FALSE, dependencies = NA)

Arguments

pkg
package description, can be path or package name. See as.package for more information

reload
if TRUE (the default), will automatically reload the package after installing.

quick
if TRUE skips docs, multiple-architectures, demos, and vignettes, to make installation as fast as possible.

local
if TRUE does not build the package first, instead installing from the local package directory. This may result in build artefacts in your package directory, but is considerably faster, and does not require a recompile every time you run it.

args
An optional character vector of additional command line arguments to bew passed to R CMD install. This defaults to the value of the option "devtools.install.args".

quiet
if TRUE suppresses output from this function.

dependencies
logical indicating to also install uninstalled packages which this pkg depends on/links to/suggests. See argument dependencies of install.packages.

Details

By default, installation takes place using the current package directory. If you have compiled code, this means that artefacts of compilation will be created in the src/ directory. If you want to avoid this, you can use local = FALSE to first build a package bundle and then install it from a temporary directory. This is slower, but keeps the source directory pristine.

If the package is loaded, it will be reloaded after installation. This is not always completely possible, see reload for caveats.

See Also

with_debug to install packages with debugging flags set.

Other package installation: install_bitbucket, install_git, install_github, install_gitorious, install_url, install_version

[Package devtools version 1.3 Index]