Install a package directly from bitbucket

This function is vectorised so you can install multiple packages in a single command.

install_bitbucket(repo, username, ref = "master", subdir = NULL,
  quiet = FALSE, auth_user = NULL, password = NULL, ...)

Arguments

repo
Repository address in the format username/repo[/subdir][@ref|#pull]. Alternatively, you can specify subdir and/or ref using the respective parameters (see below); if both are specified, the values in repo take precedence.

username
User name. Deprecated: please include username in the repo instead.

ref
Desired git reference; could be a commit, tag, or branch name. Defaults to master.

subdir
subdirectory within repo that contains the R package. Deprecated: please include subdir in repo instead.

quiet
if TRUE suppresses output from this function.

auth_user
your account username if you're attempting to install a package hosted in a private repository (and your username is different to username)

password
your password

...
Other arguments passed on to install.

See also

Examples

# NOT RUN {
install_bitbucket("sulab/mygene.r@default")
install_bitbucket("dannavarro/lsr-package")
# }