install_bitbucket {devtools}	R Documentation

Install a package directly from bitbucket

Description

This function is vectorised so you can install multiple packages in a single command.

Usage

  install_bitbucket(repo, username, ref = "master",
    branch = NULL, auth_user = NULL, password = NULL, ...)

Arguments

username
bitbucket username

repo
Repo name

ref
Desired git reference. Could be a commit, tag, or branch name. Defaults to "master".

branch
Deprecated. Use ref instead.

auth_user
your account username if you're attempting to install a package hosted in a private repository (and your username is different to username)

password
your password

...
Other arguments passed on to install.

See Also

Bitbucket API docs: https://confluence.atlassian.com/display/BITBUCKET/Use+the+Bitbucket+REST+APIs

Other package installation: install, install_git, install_github, install_gitorious, install_url, install_version

Examples

## Not run: 
install_bitbucket("paulhiemstra")
install_bitbucket(c("testrepo", "testrepo2"))

## End(Not run)

[Package devtools version 1.3 Index]