install_git {devtools}	R Documentation

Install a package from a git repository

Description

This function requires git to be installed on your system in order to be used.

Usage

  install_git(git_url, name = NULL, subdir = NULL,
    branch = NULL, git_args = character(0),
    git_binary = NULL, ...)

Arguments

branch
Name of branch or tag to use, if not master.

...
Other arguments passed on to install

git_url
Location of package. The url should point to a public or private repository.

name
Optional package name, used to provide more informative messages.

subdir
A sub-directory withing a git repository that may contain the package we are interested in installing.

git_args
A character vector providing extra arguments to pass on to

git_binary
A custom git-binary to use instead of default system's git version.

Details

It is vectorised so you can install multiple packages with a single command.

See Also

Other package installation: install, install_bitbucket, install_github, install_gitorious, install_url, install_version

Examples

## Not run: 
install_git("git://github.com/hadley/stringr.git")
install_git("git://github.com/hadley/stringr.git", branch = "stringr-0.2")

## End(Not run)

[Package devtools version 1.3 Index]