lint {devtools}	R Documentation

Lint all source files in a package.

Description

The default linters correspond to the style guide at https://style.tidyverse.org/, however it is possible to override any or all of them using the linters parameter.

Usage

lint(pkg = ".", cache = TRUE, ...)

Arguments

pkg
The package to use, can be a file path to the package or a package object. See as.package() for more information.

cache
store the lint results so repeated lints of the same content use the previous results.

...
additional arguments passed to lintr::lint_package()

Details

The lintr cache is by default stored in ~/.R/lintr_cache/ (this can be configured by setting options(lintr.cache_directory)). It can be cleared by calling lintr::clear_cache().

See Also

lintr::lint_package(), lintr::lint()

[Package devtools version 2.4.2 Index]