Release package to CRAN.

Run automated and manual tests, then ftp to CRAN.

release(pkg = ".", check = FALSE, args = NULL)

Arguments

pkg
package description, can be path or package name. See as.package for more information

check
if TRUE, run checking, otherwise omit it. This is useful if you've just checked your package and you're ready to release it.

args
An optional character vector of additional command line arguments to be passed to R CMD build.

Details

The package release process will:

Confirm that the package passes R CMD check on relevant platoforms

Confirm that important files are up-to-date

Build the package

Submit the package to CRAN, using comments in "cran-comments.md"

You can add arbitrary extra questions by defining an (un-exported) function called release_questions() that returns a character vector of additional questions to ask.

You also need to read the CRAN repository policy at https://cran.r-project.org/web/packages/policies.html and make sure you're in line with the policies. release tries to automate as many of polices as possible, but it's impossible to be completely comprehensive, and they do change in between releases of devtools.

Guarantee

If a devtools bug causes one of the CRAN maintainers to treat you impolitely, I will personally send you a handwritten apology note. Please forward me the email and your address, and I'll get a card in the mail.