reload {devtools}	R Documentation

Unload and reload package.

Description

This attempts to unload and reload an installed package. If the package is not loaded already, it does nothing. It's not always possible to cleanly unload a package: see the caveats in unload() for some of the potential failure points. If in doubt, restart R and reload the package with library().

Usage

reload(pkg = ".", quiet = FALSE)

Arguments

pkg
The package to use, can be a file path to the package or a package object. See as.package() for more information.

quiet
if TRUE suppresses output from this function.

See Also

load_all() to load a package for interactive development.

Examples

## Not run: 
# Reload package that is in current directory
reload(".")

# Reload package that is in ./ggplot2/
reload("ggplot2/")

# Can use inst() to find the package path
# This will reload the installed ggplot2 package
reload(pkgload::inst("ggplot2"))

## End(Not run)

[Package devtools version 2.4.2 Index]