Reverse dependency tools.

Tools to check and notify maintainers of all CRAN and bioconductor packages that depend on the specified package.

revdep(pkg, dependencies = c("Depends", "Imports", "Suggests", "LinkingTo"),
  recursive = FALSE, ignore = NULL, bioconductor = FALSE)

revdep_maintainers(pkg = ".")

Arguments

pkg
Package name. This is unlike most devtools packages which take a path because you might want to determine dependencies for a package that you don't have installed. If omitted, defaults to the name of the current package.

dependencies
A character vector listing the types of dependencies to follow.

recursive
If TRUE look for full set of recursive dependencies.

ignore
A character vector of package names to ignore. These packages will not appear in returned vector. This is used in revdep_check to avoid packages with installation problems or extremely long check times.

bioconductor
If TRUE also look for dependencies amongst bioconductor packages.

Details

The first run in a session will be time-consuming because it must download all package metadata from CRAN and bioconductor. Subsequent runs will be faster.

See also

Examples

# NOT RUN {
revdep("ggplot2")

revdep("ggplot2", ignore = c("xkcd", "zoo"))
# }