run_examples {devtools}	R Documentation

Run all examples in a package.

Description

One of the most frustrating parts of R CMD check is getting all of your examples to pass - whenever one fails you need to fix the problem and then restart the whole process. This function makes it a little easier by making it possible to run all examples from an R function.

Usage

run_examples(pkg = ".", start = NULL, show = TRUE, test = FALSE,
  run = TRUE, fresh = FALSE, document = TRUE)

Arguments

pkg
package description, can be path or package name. See as.package() for more information

start
Where to start running the examples: this can either be the name of Rd file to start with (with or without extensions), or a topic name. If omitted, will start with the (lexicographically) first file. This is useful if you have a lot of examples and don't want to rerun them every time you fix a problem.

show
DEPRECATED.

test
if TRUE, code in \donttest{} will be commented out. If FALSE, code in \testonly{} will be commented out.

run
if TRUE, code in \dontrun{} will be commented out.

fresh
if TRUE, will be run in a fresh R session. This has the advantage that there's no way the examples can depend on anything in the current session, but interactive code (like browser()) won't work.

document
if TRUE, document() will be run to ensure examples are updated before running them.

[Package devtools version 2.0.2 Index]