Print session information

This is sessionInfo() re-written from scratch to both exclude data that's rarely useful (e.g., the full collate string or base packages loaded) and include stuff you'd like to know (e.g., where a package was installed from).

session_info(pkgs = NULL, include_base = FALSE)

Arguments

pkgs
Either a vector of package names or NULL. If NULL, displays all loaded packages. If a character vector, also, includes all dependencies of the package.

include_base
Include base packages in summary? By default this is false since base packages should always match the R version.

Examples

session_info()

#> ─ Session info ───────────────────────────────────────────────────────────────
#>  setting  value                       
#>  version  R version 3.5.1 (2018-07-02)
#>  os       macOS Sierra 10.12.6        
#>  system   x86_64, darwin15.6.0        
#>  ui       X11                         
#>  language (EN)                        
#>  collate  en_US.UTF-8                 
#>  tz       America/New_York            
#>  date     2018-07-30                  
#> 
#> ─ Packages ───────────────────────────────────────────────────────────────────
#>  package     * version     date       source                          
#>  assertthat    0.2.0       2017-04-11 CRAN (R 3.5.0)                  
#>  backports     1.1.2       2017-12-13 CRAN (R 3.5.0)                  
#>  callr         2.0.4       2018-05-15 cran (@2.0.4)                   
#>  cli           1.0.0       2017-11-05 CRAN (R 3.5.0)                  
#>  clisymbols    1.2.0       2017-05-21 cran (@1.2.0)                   
#>  commonmark    1.5         2018-04-28 cran (@1.5)                     
#>  crayon        1.3.4       2017-09-16 CRAN (R 3.5.0)                  
#>  curl          3.2         2018-03-28 cran (@3.2)                     
#>  desc          1.2.0       2018-05-01 cran (@1.2.0)                   
#>  devtools    * 1.13.6.9000 <NA>       local                           
#>  digest        0.6.15      2018-01-28 CRAN (R 3.5.0)                  
#>  evaluate      0.11        2018-07-17 cran (@0.11)                    
#>  fs            1.2.3       2018-06-08 cran (@1.2.3)                   
#>  git2r         0.23.0.9000 2018-07-30 Github (ropensci/git2r@822155b) 
#>  glue          1.3.0       2018-07-27 Github (tidyverse/glue@a292148) 
#>  highlight     0.4.7.2     2017-10-04 cran (@0.4.7.2)                 
#>  htmltools     0.3.6       2017-04-28 CRAN (R 3.5.0)                  
#>  httr          1.3.1       2017-08-20 CRAN (R 3.5.0)                  
#>  jsonlite      1.5         2017-06-01 CRAN (R 3.5.0)                  
#>  knitr         1.20        2018-02-20 CRAN (R 3.5.0)                  
#>  magrittr      1.5         2014-11-22 CRAN (R 3.5.0)                  
#>  MASS          7.3-50      2018-04-30 CRAN (R 3.5.1)                  
#>  memoise       1.1.0       2018-07-26 Github (hadley/memoise@06d16ec) 
#>  nvimcom     * 0.9-48      2018-07-27 local                           
#>  pillar        1.3.0       2018-07-14 cran (@1.3.0)                   
#>  pkgbuild      1.0.0       2018-06-27 cran (@1.0.0)                   
#>  pkgdown       1.1.0       2018-06-02 cran (@1.1.0)                   
#>  pkgload       1.0.0       2018-07-07 cran (@1.0.0)                   
#>  prettyunits   1.0.2       2015-07-13 CRAN (R 3.5.0)                  
#>  processx      3.1.0       2018-05-15 cran (@3.1.0)                   
#>  purrr         0.2.5       2018-05-29 cran (@0.2.5)                   
#>  R6            2.2.2       2017-06-17 CRAN (R 3.5.0)                  
#>  rcmdcheck     1.2.1.9005  2018-07-26 Github (r-lib/rcmdcheck@ad42561)
#>  Rcpp          0.12.18     2018-07-23 cran (@0.12.18)                 
#>  rematch       1.0.1       2016-04-21 cran (@1.0.1)                   
#>  rlang         0.2.1       2018-05-30 cran (@0.2.1)                   
#>  rmarkdown     1.10        2018-06-11 cran (@1.10)                    
#>  roxygen2      6.0.1       2017-02-06 CRAN (R 3.5.0)                  
#>  rprojroot     1.3-2       2018-01-03 cran (@1.3-2)                   
#>  rstudioapi    0.7         2017-09-07 CRAN (R 3.5.0)                  
#>  sessioninfo   1.0.0       2017-06-21 cran (@1.0.0)                   
#>  stringi       1.2.4       2018-07-20 cran (@1.2.4)                   
#>  stringr       1.3.1       2018-05-10 CRAN (R 3.5.0)                  
#>  testthat    * 2.0.0       2017-12-13 CRAN (R 3.5.0)                  
#>  tibble        1.4.2       2018-01-22 cran (@1.4.2)                   
#>  usethis     * 1.3.0.9000  2018-07-27 Github (r-lib/usethis@585b9f7)  
#>  whisker       0.3-2       2013-04-28 CRAN (R 3.5.0)                  
#>  withr         2.1.2       2018-03-15 CRAN (R 3.5.0)                  
#>  xml2          1.2.0       2018-01-24 cran (@1.2.0)                   
#>  yaml          2.2.0       2018-07-25 cran (@2.2.0)                   

session_info("devtools")

#> ─ Session info ───────────────────────────────────────────────────────────────
#>  setting  value                       
#>  version  R version 3.5.1 (2018-07-02)
#>  os       macOS Sierra 10.12.6        
#>  system   x86_64, darwin15.6.0        
#>  ui       X11                         
#>  language (EN)                        
#>  collate  en_US.UTF-8                 
#>  tz       America/New_York            
#>  date     2018-07-30                  
#> 
#> ─ Packages ───────────────────────────────────────────────────────────────────
#>  package     * version     date       source                          
#>  assertthat    0.2.0       2017-04-11 CRAN (R 3.5.0)                  
#>  backports     1.1.2       2017-12-13 CRAN (R 3.5.0)                  
#>  callr         2.0.4       2018-05-15 cran (@2.0.4)                   
#>  cli           1.0.0       2017-11-05 CRAN (R 3.5.0)                  
#>  clipr         0.4.1       2018-06-23 cran (@0.4.1)                   
#>  clisymbols    1.2.0       2017-05-21 cran (@1.2.0)                   
#>  crayon        1.3.4       2017-09-16 CRAN (R 3.5.0)                  
#>  curl          3.2         2018-03-28 cran (@3.2)                     
#>  desc          1.2.0       2018-05-01 cran (@1.2.0)                   
#>  devtools    * 1.13.6.9000 <NA>       local                           
#>  digest        0.6.15      2018-01-28 CRAN (R 3.5.0)                  
#>  fansi         0.2.3       2018-05-06 cran (@0.2.3)                   
#>  fs            1.2.3       2018-06-08 cran (@1.2.3)                   
#>  gh            1.0.1       2018-07-30 github (gaborcsardi/gh)         
#>  git2r         0.23.0.9000 2018-07-30 Github (ropensci/git2r@822155b) 
#>  glue          1.3.0       2018-07-27 Github (tidyverse/glue@a292148) 
#>  httr          1.3.1       2017-08-20 CRAN (R 3.5.0)                  
#>  ini           0.3.1       2018-05-20 cran (@0.3.1)                   
#>  jsonlite      1.5         2017-06-01 CRAN (R 3.5.0)                  
#>  magrittr      1.5         2014-11-22 CRAN (R 3.5.0)                  
#>  memoise       1.1.0       2018-07-26 Github (hadley/memoise@06d16ec) 
#>  mime          0.5         2016-07-07 CRAN (R 3.5.0)                  
#>  openssl       1.0.1       2018-03-03 cran (@1.0.1)                   
#>  pillar        1.3.0       2018-07-14 cran (@1.3.0)                   
#>  pkgbuild      1.0.0       2018-06-27 cran (@1.0.0)                   
#>  pkgload       1.0.0       2018-07-07 cran (@1.0.0)                   
#>  praise        1.0.0       2015-08-11 CRAN (R 3.5.0)                  
#>  prettyunits   1.0.2       2015-07-13 CRAN (R 3.5.0)                  
#>  processx      3.1.0       2018-05-15 cran (@3.1.0)                   
#>  R6            2.2.2       2017-06-17 CRAN (R 3.5.0)                  
#>  rcmdcheck     1.2.1.9005  2018-07-26 Github (r-lib/rcmdcheck@ad42561)
#>  Rcpp          0.12.18     2018-07-23 cran (@0.12.18)                 
#>  rematch2      2.0.1       2017-06-20 cran (@2.0.1)                   
#>  rlang         0.2.1       2018-05-30 cran (@0.2.1)                   
#>  rprojroot     1.3-2       2018-01-03 cran (@1.3-2)                   
#>  rstudioapi    0.7         2017-09-07 CRAN (R 3.5.0)                  
#>  sessioninfo   1.0.0       2017-06-21 cran (@1.0.0)                   
#>  testthat    * 2.0.0       2017-12-13 CRAN (R 3.5.0)                  
#>  tibble        1.4.2       2018-01-22 cran (@1.4.2)                   
#>  usethis     * 1.3.0.9000  2018-07-27 Github (r-lib/usethis@585b9f7)  
#>  utf8          1.1.4       2018-05-24 cran (@1.1.4)                   
#>  whisker       0.3-2       2013-04-28 CRAN (R 3.5.0)                  
#>  withr         2.1.2       2018-03-15 CRAN (R 3.5.0)