submit_cran {devtools}	R Documentation

Submit a package to CRAN.

Description

This uses the new CRAN web-form submission process. After submission, you will receive an email asking you to confirm submission - this is used to check that the package is submitted by the maintainer.

Usage

submit_cran(pkg = ".", args = NULL)

Arguments

pkg
The package to use, can be a file path to the package or a package object. See as.package() for more information.

args
An optional character vector of additional command line arguments to be passed to R CMD build.

Details

It's recommended that you use release() rather than this function as it performs more checks prior to submission.

[Package devtools version 2.4.2 Index]