Submit a package to CRAN

This submits your package to CRAN using the web-form submission process. After submission, you will receive an email asking you to confirm submission

this is used to check that the package is submitted by the maintainer.

You may prefer to use submit_cran() indirectly, by calling release() instead. release() performs many checks verifying that your package is indeed ready for CRAN, before eventually asking for your confirmation that you'd like to submit it to CRAN (which it does by calling submit_cran()).

Whether to use release() or submit_cran() depends on the rest of your development process. If you want to be super cautious, use release(), even though it may be redundant with other checks you have performed. On the other hand, if you have many other checks in place (such as automated checks via GitHub Actions and the task list generated by usethis::use_release_issue()), it makes sense to use submit_cran() directly.

Usage

submit_cran(pkg = ".", args = NULL)

Arguments

pkg

The package to use, can be a file path to the package or a package object. See as.package() for more information.

args

An optional character vector of additional command line arguments to be passed to R CMD build.