Execute test_that tests in a package.

test() is a shortcut for testthat::test_dir(), it runs all of a package's tests. test_file runs test() on the active file. test_coverage() computes test coverage for your package. It is a shortcut for covr::package_coverage() and covr::report(). test_coverage_file() computes test coverage for the active file. Is a shortcut for covr::file_coverage() and covr::report().

test(pkg = ".", filter = NULL, ...)

test_coverage(pkg = ".", ...)

uses_testthat(pkg = ".")

test_file(file = find_active_file(), ...)

test_coverage_file(file = find_active_file(), ...)

Arguments

pkg
package description, can be path or package name. See as.package() for more information

filter
If not NULL, only tests with file names matching this regular expression will be executed. Matching will take on the file name after it has been stripped of "test-" and ".R".

...
additional arguments passed to testthat::test_dir() and covr::package_coverage()

file
One or more source or test files. If a source file the corresponding test file will be run. The default is to use the active file in RStudio (if available).