wd {devtools}	R Documentation

Set working directory.

Description

Set working directory.

Usage

wd(pkg = ".", path = "")

Arguments

pkg
The package to use, can be a file path to the package or a package object. See as.package() for more information.

path
path within package. Leave empty to change working directory to package directory.

[Package devtools version 2.4.2 Index]